+++ .gitignore
@@ -4,3 +4,4 @@ res/nikita/*
 *.code-workspace
 node_modules
 sandbox
+srvSpiralSection
